+++ R/backend-teradata.R
@@ -32,6 +32,7 @@ dbplyr_edition.Teradata <- function(con) {
 #' @export
 sql_query_select.Teradata <- function(con, select, from, where = NULL,
                                              group_by = NULL, having = NULL,
+                                             window = NULL,
                                              order_by = NULL,
                                              limit = NULL,
                                              distinct = FALSE,
@@ -45,6 +46,7 @@ sql_query_select.Teradata <- function(con, select, from, where = NULL,
     where     = sql_clause_where(where),
     group_by  = sql_clause_group_by(group_by),
     having    = sql_clause_having(having),
+    window    = sql_clause_window(window),
     order_by  = sql_clause_order_by(order_by, subquery, limit),
     lvl       = lvl
   )