+++ NEWS.md
@@ -1,5 +1,9 @@
 # dbplyr (development version)
 
+* `mutate()` and `transmute()` now automatically use named windows if a window
+  definition is used at least twice and the backend supports named windows
+  (@mgirlich, #624).
+
 * `across()` now works if passing the name of a list of functions to the `.fns`
   argument (@mgirlich, #817).